+++ scitos_behavior_tree/include/scitos_behavior_tree/condition/is_bumper_activated_condition.hpp
@@ -52,7 +52,7 @@ class IsBumperActivatedCondition : public BT::ConditionNode{
 		static BT::PortsList providedPorts()	{
 			return {
 			BT::InputPort<std::string>(
-				"bumper", std::string("/bumper"), "Bumper topic")
+				"bumper_topic", std::string("/bumper"), "Bumper topic")
 			};
 		}