+++ scitos_mira/include/scitos_mira/ScitosModule.hpp
@@ -1,7 +1,7 @@
 /*
  * SCITOS MODULE
  *
- * Copyright (c) 2022 Alberto José Tudela Roldán <ajtudela@gmail.com>
+ * Copyright (c) 2022-2023 Alberto José Tudela Roldán <ajtudela@gmail.com>
  * 
  * This file is part of scitos_mira project.
  * 
@@ -20,15 +20,17 @@
 
 // ROS
 #include "rclcpp/rclcpp.hpp"
+#include "rclcpp_lifecycle/lifecycle_node.hpp"
 
 /**
  * @brief Base class for all Scitos modules (Drive, Charger, etc).
  * 
  */
-class ScitosModule : public rclcpp::Node{
+class ScitosModule{
 	public:
-		ScitosModule(const std::string& name);
-		virtual void initialize() = 0;
+		ScitosModule(const std::string & name);
+		virtual void configure(const rclcpp_lifecycle::LifecycleNode::WeakPtr & ros_node) = 0;
+		virtual void reset_publishers() = 0;
 		virtual ~ScitosModule(){};
 
 	protected:
@@ -37,7 +39,8 @@ class ScitosModule : public rclcpp::Node{
 		std::string get_mira_param(std::string param_name);
 
 		mira::Authority authority_;
+		rclcpp_lifecycle::LifecycleNode::WeakPtr node_;
+		rclcpp::Logger logger_;
 };
 
 #endif // SCITOS_MIRA__SCITOS_MODULE_HPP_
-