+++ scitos_mira/include/scitos_mira/modules/ScitosCharger.hpp
@@ -1,7 +1,7 @@
 /*
  * SCITOS CHARGER
  *
- * Copyright (c) 2022 Alberto José Tudela Roldán <ajtudela@gmail.com>
+ * Copyright (c) 2022-2023 Alberto José Tudela Roldán <ajtudela@gmail.com>
  * 
  * This file is part of scitos_mira project.
  * 
@@ -17,6 +17,8 @@
 
 // ROS
 #include "rclcpp/rclcpp.hpp"
+#include "rclcpp_lifecycle/lifecycle_node.hpp"
+#include "rclcpp_lifecycle/lifecycle_publisher.hpp"
 #include "sensor_msgs/msg/battery_state.hpp"
 
 // SCITOS MSGS
@@ -35,22 +37,24 @@ class ScitosCharger : public ScitosModule{
 			return std::shared_ptr<ScitosModule>(new ScitosCharger());
 		}
 
-		void initialize();
+		void configure(const rclcpp_lifecycle::LifecycleNode::WeakPtr & ros_node);
+		void reset_publishers();
 
 	private:
-		rclcpp::Publisher<sensor_msgs::msg::BatteryState>::SharedPtr battery_pub_;
-		rclcpp::Publisher<scitos_msgs::msg::ChargerStatus>::SharedPtr charger_pub_;
+		rclcpp_lifecycle::LifecyclePublisher<sensor_msgs::msg::BatteryState>::SharedPtr battery_pub_;
+		rclcpp_lifecycle::LifecyclePublisher<scitos_msgs::msg::ChargerStatus>::SharedPtr charger_pub_;
 		
-		rclcpp::Service<scitos_msgs::srv::SavePersistentErrors>::SharedPtr save_persistent_errors_service_;
+		rclcpp::Service<scitos_msgs::srv::SavePersistentErrors>::SharedPtr 
+			save_persistent_errors_service_;
 		
 		ScitosCharger();
 
 		void battery_data_callback(mira::ChannelRead<mira::robot::BatteryState> data);
 		void charger_status_callback(mira::ChannelRead<uint8> data);
 
-		bool save_persistent_errors(const std::shared_ptr<scitos_msgs::srv::SavePersistentErrors::Request> request,
-								std::shared_ptr<scitos_msgs::srv::SavePersistentErrors::Response> response);
+		bool save_persistent_errors(
+			const std::shared_ptr<scitos_msgs::srv::SavePersistentErrors::Request> request,
+			std::shared_ptr<scitos_msgs::srv::SavePersistentErrors::Response> response);
 };
 
 #endif // SCITOS_MIRA__SCITOS_CHARGER_HPP_
-