+++ scitos_mira/include/scitos_mira/modules/ScitosEBC.hpp
@@ -1,7 +1,7 @@
 /*
  * SCITOS EBC
  *
- * Copyright (c) 2022 Alberto José Tudela Roldán <ajtudela@gmail.com>
+ * Copyright (c) 2022-2023 Alberto José Tudela Roldán <ajtudela@gmail.com>
  * 
  * This file is part of scitos_mira project.
  * 
@@ -14,6 +14,7 @@
 
 // ROS
 #include "rclcpp/rclcpp.hpp"
+#include "rclcpp_lifecycle/lifecycle_node.hpp"
 #include "rcl_interfaces/msg/set_parameters_result.hpp"
 
 #include "scitos_mira/ScitosModule.hpp"
@@ -28,16 +29,17 @@ class ScitosEBC : public ScitosModule{
 			return std::shared_ptr<ScitosModule>(new ScitosEBC());
 		}
 
-		void initialize();
+		void configure(const rclcpp_lifecycle::LifecycleNode::WeakPtr & ros_node);
+		void reset_publishers();
 
 	private:
 		ScitosEBC();
 
-		OnSetParametersCallbackHandle::SharedPtr callback_handle_;
+		rclcpp::node_interfaces::OnSetParametersCallbackHandle::SharedPtr dyn_params_handler_;
 
-		rcl_interfaces::msg::SetParametersResult parameters_callback(const std::vector<rclcpp::Parameter> &parameters);
+		rcl_interfaces::msg::SetParametersResult parameters_callback(
+			const std::vector<rclcpp::Parameter> &parameters);
 
 };
 
 #endif // SCITOS_MIRA__SCITOS_EBC_HPP_
-