+++ scitos_mira/src/ScitosMira.cpp
@@ -1,7 +1,7 @@
 /*
  * SCITOS MIRA
  *
- * Copyright (c) 2022 Alberto José Tudela Roldán <ajtudela@gmail.com>
+ * Copyright (c) 2022-2023 Alberto José Tudela Roldán <ajtudela@gmail.com>
  * 
  * This file is part of scitos_mira project.
  * 
@@ -16,39 +16,51 @@
 // BOOST
 #include <boost/algorithm/string/join.hpp>
 
+// ROS
+#include "nav2_util/node_utils.hpp"
+
 #include "scitos_mira/ScitosMira.hpp"
 #include "scitos_mira/ModuleFactory.hpp"
 
-ScitosMira::ScitosMira(const std::string& name) : Node(name), framework_(args_){
+ScitosMira::ScitosMira(const std::string& name, bool intra_process_comms) : 
+					rclcpp_lifecycle::LifecycleNode(name, rclcpp::NodeOptions()
+						.use_intra_process_comms(intra_process_comms)), 
+					framework_({}){
+}
+
+rclcpp_CallReturn ScitosMira::on_configure(const rclcpp_lifecycle::State &){
 	RCLCPP_INFO(this->get_logger(), "Configuring the node...");
 
 	// Redirect Mira logger
 	MIRA_LOGGER.registerSink(RosLogSink(this->get_logger()));
 
 	// Declare and read parameters
-	declare_parameter_if_not_declared("modules", rclcpp::ParameterType::PARAMETER_STRING_ARRAY, 
-							rcl_interfaces::msg::ParameterDescriptor()
-							.set__description("List of the modules exposed by the node"));
+	nav2_util::declare_parameter_if_not_declared(this, "modules", 
+		rclcpp::ParameterValue(std::vector<std::string>{}), 
+		rcl_interfaces::msg::ParameterDescriptor()
+			.set__description("List of the modules exposed by the node"));
 	this->get_parameter("modules", modules_names_);
 	std::string joined = boost::algorithm::join(modules_names_, ", ");
 	if (!joined.empty()){
 		RCLCPP_INFO(this->get_logger(), "Loaded modules: [%s]", joined.c_str());
 	}else{
-		RCLCPP_ERROR(this->get_logger(), "Can't read parameter 'modules'. This MUST be supplied as a space separated list of SCITOS hardware modules to interface into ROS");
-		exit(1);
+		RCLCPP_ERROR_STREAM(this->get_logger(), "Can't read parameter 'modules'. " 
+			<< "This MUST be supplied as a space separated list of SCITOS "
+			<< "hardware modules to interface into ROS");
+		return rclcpp_CallReturn::FAILURE;
 	}
 
 	std::string config;
-	declare_parameter_if_not_declared("scitos_config", rclcpp::ParameterType::PARAMETER_STRING, 
-							rcl_interfaces::msg::ParameterDescriptor()
-							.set__description("Configuration of the robot in XML format"));
+	nav2_util::declare_parameter_if_not_declared(this, "scitos_config", rclcpp::ParameterValue(""), 
+		rcl_interfaces::msg::ParameterDescriptor()
+			.set__description("Configuration of the robot in XML format"));
 	this->get_parameter("scitos_config", config);
 	if (!config.empty()){
 		RCLCPP_INFO(this->get_logger(), "Loaded scitos config: %s", config.c_str());
 		framework_.load(config);
 	}else{
 		RCLCPP_ERROR(this->get_logger(), "Can't read parameter 'scitos_config'");
-		exit(1);
+		return rclcpp_CallReturn::FAILURE;
 	}
 
 	// Sleep for 2 seconds and start Mira framework
@@ -59,31 +71,70 @@ ScitosMira::ScitosMira(const std::string& name) : Node(name), framework_(args_){
 	ModuleFactory& factory = ModuleFactory::get_instance();
 	for (const auto& name : modules_names_){
 		if (!factory.is_registered(name)){
-			RCLCPP_ERROR_STREAM(this->get_logger(), "A non existent module named '" << name << "' was trying to be created.");
+			RCLCPP_ERROR_STREAM(this->get_logger(), "A non existent module named '" 
+				<< name << "' was trying to be created.");
+			return rclcpp_CallReturn::ERROR;
 		}else{
 			RCLCPP_INFO(this->get_logger(), "Created module: [%s]", name.c_str());
 			modules_.push_back(factory.create_module(name));
 		}
 	}
 
-	initialize();
+	// Configure MIRA modules
+	configure_modules();
+
+	return rclcpp_CallReturn::SUCCESS;
 }
 
 ScitosMira::~ScitosMira(){
-}
-
-void ScitosMira::initialize(){
-	// Initialize all of the modules
-	for (auto& module: modules_){
-		module->initialize();
+	if (framework_.isTerminationRequested()){
+		RCLCPP_INFO(this->get_logger(), "Stopping MIRA framework...");
 	}
 }
 
-void ScitosMira::declare_parameter_if_not_declared(const std::string & param_name, 
-	const rclcpp::ParameterType & param_type,
-	const rcl_interfaces::msg::ParameterDescriptor & param_descriptor){
+rclcpp_CallReturn ScitosMira::on_activate(const rclcpp_lifecycle::State & state){
+	LifecycleNode::on_activate(state);
+	RCLCPP_INFO(this->get_logger(), "Activating the node...");
 
-	if (!this->has_parameter(param_name)){
-		this->declare_parameter(param_name, param_type, param_descriptor);
+	return rclcpp_CallReturn::SUCCESS;
+}
+
+rclcpp_CallReturn ScitosMira::on_deactivate(const rclcpp_lifecycle::State & state){
+	LifecycleNode::on_deactivate(state);
+	RCLCPP_INFO(this->get_logger(), "Deactivating the node...");
+
+	return rclcpp_CallReturn::SUCCESS;
+}
+
+rclcpp_CallReturn ScitosMira::on_cleanup(const rclcpp_lifecycle::State &){
+	RCLCPP_INFO(this->get_logger(), "Cleaning the node...");
+
+	// Release the shared pointers
+	reset_publishers_modules();
+
+	return rclcpp_CallReturn::SUCCESS;
+}
+
+rclcpp_CallReturn ScitosMira::on_shutdown(const rclcpp_lifecycle::State & state){
+	RCLCPP_INFO(this->get_logger(), "Shutdown the node from state %s.", state.label().c_str());
+
+	// Release the shared pointers
+	reset_publishers_modules();
+
+	return rclcpp_CallReturn::SUCCESS;
+}
+
+void ScitosMira::configure_modules(){
+	// Configure all modules
+	auto node = this->shared_from_this();
+	for (auto& module: modules_){
+		module->configure(node);
+	}
+}
+
+void ScitosMira::reset_publishers_modules(){
+	// Reset the publishers of all modules
+	for (auto& module: modules_){
+		module->reset_publishers();
 	}
 }